Bugfix: Let packet_socket_receive honor the direction when using libpcap.
When using libpcap, the function packet_socket_receive was ignoring the
direction. If there were inbound packets sniffed the following try to
read the packet from the tun descriptor blocked.
This patch makes packet_socket_receive aware of the direction and therefore
the libpcap implementation has the same functionality as the Linux one.
It just uses two pcap handles, one for each direction.

Tested on FreeBSD.

Signed-off-by: Michael Tuexen <tuexen@fh-muenster.de>