Fix alignment of headers in the packet buffer
This patch makes sure that all heaaders start at a four byte
boundary. This is done by not having the ethernet header, which
is 14 bytes long, in the packet buffer.
This fixes the remote mode for platforms having strict alignment
restrictions, like ARM. Without this patch you can't use the remote
mode on a Rasperry Pi running FreeBSD.
This patch has been tested on Linux (Intel) in local and remote mode
and one FreeBSD (Intel and ARM) in local and remote mode.

Signed-off-by: Michael Tuexen <tuexen@fh-muenster.de>