Fix setting the MTU and speed via the command line
For FreeBSD use ioctl() to set the speed and mtu of the tun
interface instead of the Linux specific ethtool for the speed.
Also ensure that you can set the MTU to the default value in
case it is configured to a different one.
On non-FreeBSD platforms restore the old behaviour.
This fixes https://github.com/nplab/packetdrill/issues/17