Allow automatic tun device selection on *BSD.
When using local mode on *BSD, using the tun0 device was hard
coded in contrast to Linux, where the next unused tun device
was used.
This patch changes the behaviour on *BSD to use the next unused
tun device. This also requires that the device is removed when
packetdrill terminates. This was the case on Linux, but not on
*BSD.
On *BSD there is also a new command line option (--tun-dev) to
set the device name and not select an ephemerial one.