Add an option to keep the tun interface up on *BSD.
On *BSD platforms the original version always used the tun0 device
in local mode and never deleted it, except on NetBSD. This was
handy when using for example Wireshark for capturing trafic.
The last commit removed this possibility, since it removes the
interface. Add the --persistent_tun_dev command line option
to bring it back.

To get the behaviour of the original version, use on *BSD
platforms --tun-dev=tun0 --persistent_tun_dev.